requirements.txt
psycopg2-binary==2.9.9
requests==2.25.1
pandas==  2.1.1
openpyxl==3.1.2
python-dotenv==1.0.0
SQLAlchemy==2.0.21